// greek-mythos-vue/src/models/app-user.ts
export interface AppUser {
    uid: string;
    displayName: string;
    createdAt: number | Object;
    lastOnline: number | Object;
    email: string;
    active: boolean;
    imageUrl?: string;
    gamesPlayed?: number;
    gamesWon?: number;
    inGame?: {
        gameId: string;
        timestamp: number;
    }
}